Ensure close file descriptors by using context manager in test suite.
Generally, it is true that as file descriptors are closed when their
underlying objects are deleted, we don't need close them explicitly.
However if a Python language implementation that uses other strategy
than reference count model, there is no warranty it will happen
immediately after when those objects lose the last reference.  So we
use context manager to ensure close() is called immediately after
those objects to be unnecessary.  This helps us to run 'make check'
with PyPy even smaler limit of number of open files, although we
don't recommend use it because it is much slower than CPython for
this purpose.

* build/run_tests.py
  (TestHarness._init_c_tests): Use context manager to write file.
  (TestHarness.Job.execute,
  TestHarness.CollectingThread._count_c_tests,
  Testharness.CollectingThread._count_py_tests,):
    Use context manager for Popen object.

* subversion/tests/cmdline/svntest/actions.py
  (load_repo, load_dumpfile): Use context manager to read file contents.

* subversion/tests/cmdline/svntest/main.py
  (trust_ssl_cert): Use context manager to read file contents.
  
* subversion/tests/cmdline/svntest/sandbox.py
  (Sandbox._ensure_authz get_content): New function.
  (Sandbox._ensure_authz): Use it.

* subversion/tests/cmdline/svntest/tree.py
  (get_text): Use context manager to read file contents.

* subversion/tests/cmdline/svntest/wc.py
  (State.from_wc): Use context manager to read file contents.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1884279 13f79535-47bb-0310-9956-ffa450edef68