Follow up to r1880967: swig-py: Restore return value type of svn.fs.commit_txn
On r1880967 we reverted a "wrong" behavior of svn.fs.commit_txn, that it
returned always a 2-tuple of None and new_rev if it didn't raise Exception.
However we reconsidered that it is more important to keep compatibility on
existing program codes.  To keep code simple, we adjust return value on
Python code rather than SWIG typemap. 

* subversion/bindings/swig/python/svn/fs.py
  (): rename the wrapper function svn_fs_commit_txn imported from
    libsvn.fs module into internal name '_svn_fs_commit_txn'.
  (svn_fs_commit_txn):
    New function to wrap _svn_fs_commit_txn to adjust return value type.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1885007 13f79535-47bb-0310-9956-ffa450edef68