Lay a bit of necessary groundwork for streamy checkouts.
Make the get_info function for an install stream return only the parts
of the file information that we require right now (mtime and size), instead
of supporting arbitrary APR finfo requests.  This should make it easier to
introduce an ability to set custom mtime and attributes for the install
stream.

* subversion/include/private/svn_io_private.h
  (svn_stream__install_get_info): Return the mtime and size as two optional
   arguments.

* subversion/libsvn_subr/stream.c
  (svn_stream__install_get_info): Rework to make the apr_file_info_get()
   call with predefined flags, depending on which parts of the information
   the caller has asked for.

* subversion/libsvn_wc/wc_db_pristine.c
  (pristine_install_txn): Update calls to svn_stream__install_get_info().


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1886489 13f79535-47bb-0310-9956-ffa450edef68