Revert r1887111 and r1887112, which inlined the svn_subst_translation_required() check in the working file writer.
With a second thought, it might be better to keep using that helper to rely on
the stable primitive, instead of reimplementing it by hand.

And for the keywords part of that check, we should be able to get rid of it
with a bit of groundwork in subst.c.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1887172 13f79535-47bb-0310-9956-ffa450edef68