Fix issue #4874 "Avoid foreign repository merge if repository UUIDs match"
Summary:  Repository root URLs must be consistent in a merge.  A merge with
differing source and target repository root URLs pointing to the same
repository (same UUID) previously did a foreign-repository merge and is now
disallowed.  A two-URL merge with differing source URLs pointing to the same
repository (same UUID) was previously broken and is now disallowed.

The initial problem, encountered in real life, was that Subversion did a
foreign-repository merge when the repository root URLs of source and target
differed but pointed to equivalent repositories (having the same UUID).

In principle Subversion should do a normal merge, not a foreign-repository
merge.  However, for both historical reasons (the behaviour had previously
changed twice, in Subversion 1.6 and 1.8) and practical reasons (the code
base is not written to respect different URLs pointing to the same or
equivalent repositories), and because we expect this case to be unnecessary
and rarely encountered, we instead disallow this case.  Work-arounds are
noted in the issue.

We also fix a similar issue in the comparison of the two source URLs in a
two-URL merge.  In this comparison, there is no "foreign repository" case to
consider.  Previously this comparison only ensured their UUIDs matched,
which is logically correct.  However, in practice the code base requires
that the two given repository root URLs must be identical in order to ensure
that all subsequent operations on those URLs are consistent, including the
source-target comparison which in fact just compares source1 with target.

New behaviour:

  * comparing the two source repository roots in a two-URL merge:
    - same repository root URL          -> ok
    - different repository root URL     -> error

  * comparing source and target repository roots:
    - same repository root URL          -> same-repository merge
    - different URLs, different UUIDs   -> foreign-repository merge
    - different URLs, same UUID         -> error

(This does not change the fact that the foreign-repository merge case is
not available in "automatic" or "reintegrate" merges.)

* subversion/libsvn_client/merge.c
  (is_same_repos, check_same_repos): Implement the stricter checks, returning
    an error in the case of different URLs but same UUIDs. Drop the option to
    choose "strict" URL comparison.
  (svn_client__merge_locked,
   open_reintegrate_source_and_target,
   merge_peg_locked,
   client_find_automatic_merge): Use the stricter checks in all cases. Add
    context-specific error messages in all cases.

* subversion/tests/cmdline/merge_tests.py
  (merge_error_if_source_urls_differ,
   merge_error_if_target_url_differs_and_same_uuid): New tests.
  (test_list): Run them.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1888474 13f79535-47bb-0310-9956-ffa450edef68