Fix an assertion failure triggered by the conflict resolver.
The resolver may need to deal with nodes of type 'none' but was using
a token map which does not support the 'none' type. Switch to a token
map which supports the 'none' type. This makes the regression test
added in r1892470 pass.

The above fix uncovered a different problem where the wc_move_targets
pointer in conflict details for a "local missing" node was left as NULL.
The assumption throughout the resolver is that this variable is initialized
to an empty array instead. Do this to avoid a NULL-deref which would
otherwise be triggered by the same regression test.

* subversion/libsvn_client/conflicts.c
  (conflict_tree_get_details_local_missing): Always initialize the
    details->wc_move_targets variable.

* subversion/libsvn_wc/wc_db.c
  (svn_wc__db_find_working_nodes_with_basename,
   svn_wc__db_find_copies_of_repos_path): Use kind_map_none instead of
    the regular kind_map which doesn't support the 'none' type.

* subversion/tests/cmdline/merge_tree_conflict_tests.py
  (merge_local_missing_node_kind_none): Remove XFail marker.

Reported by: Joshua Kordani (jkordani {AT} roboticresearch dot com)


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1892471 13f79535-47bb-0310-9956-ffa450edef68