On the pristines-on-demand branch: Revert r1801448 ("Temporarily revert r1759233 …").
This effectively reinstates the changes made in r1759233.

The r1759233 commit started comparing checksums in the implementation of
the svn_wc__internal_file_modified_p() function.  More specifically, it
changed the approach where we simultaneously read the pristine and
working files into calculating the checksum of the working file and
comparing it against the checksum recorded in wc.db.

We can't get too far away in the attempt to make the pristine contents
optional if we still need it to understand if the file has been modified.
So let's untangle that first by switching to a checksum comparison.

* subversion/libsvn_wc/questions.c
  (compare_and_verify): Replace PRISTINE_STREAM and PRISTINE_SIZE arguments
   with PRISTINE_CHECKSUM. Calculate SHA-1 checksum of detranslated
   (normalized) content and compare it with what's stored in wc.db. Open
   pristine file only if exact comparison is requested.
  (svn_wc__internal_file_modified_p): Do not open pristine file here, just
   pass pristine's checksum to compare_and_verify().


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/pristines-on-demand@1892645 13f79535-47bb-0310-9956-ffa450edef68