On the pristines-on-demand branch: Don't read the pristine text even when checking for file modifications in the "exact comparison" mode.
Previously for that kind of comparison we'd check if the pristine text
translated according to the file props matches the current contents of
the file.  This allows detecting modifications that are otherwise missed
by the non-exact comparison where we check if the *detranslated* contents
of the file matches the pristine.  One example of a modification that is
only detected in the exact comparison mode is a file with manually
contracted svn:keywords.

To make this work without a pristine text, let's check for two things:

1) That the checksum of the detranslated contents matches the recorded
   pristine checksum, as in the case of a non-exact comparison, …

2) …and additionally, that the contents of the working file does not
   change when retranslated according to its properties.

Technically we're going to do that with a single read of the file contents,
while checksumming it's original, detranslated and retranslated versions.

* subversion/libsvn_wc/questions.c
  (compare_and_verify): Check for an unrecognized eol-style value when
   fetching the file props, allowing for an early error-out.  When asked
   to perform an exact comparison, calculate checksums for the original,
   detranslated and retranslated versions of the working file.  Compare the
   detranslated-pristine and original-retranslated checksum pairs to see
   if the file is modified.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/pristines-on-demand@1892646 13f79535-47bb-0310-9956-ffa450edef68