On the pristines-on-demand branch: Lay some more groundwork for making the text-bases optional.
For now, let's ensure we only use the explicit version of the file_install
workqueue item, where the caller specifies the install source in the form of
a file path.

(The implicit version of this work item only specifies the checksum of the
pristine text, and its contents is taken from the pristine store when the
work item is processed.  If we make the text-bases optional, supporting
this type of item would probably require additional bookkeeping, where
we'd have to ensure that the contents is still present at the moment when
we process the workqueue item.  Which should be possible, but for now let's
transition into a state where we explicitly specify the target contents,
and maybe at some point we could work our way backwards.

Note that with r1886490, we mostly use the working_file_writer to install
the files, so explicitly providing the contents here shouldn't introduce an
overhead for common checkout and update cases.)

* subversion/libsvn_wc/adm_crawler.c
  (restore_file): Setaside the text-base contents for the work item.

* subversion/libsvn_wc/copy.c
  (): Include textbase.h.
  (restore_file): Setaside the text-base contents for the work item.

* subversion/libsvn_wc/externals.c
  (close_file): Setaside the text-base contents for the work item.
   Use svn_wc__db_pristine_read() instead of the svn_wc__textbase layer,
   because we are just about to call svn_wc__db_external_add_file(), so at
   this point our file external does not yet exist.

* subversion/libsvn_wc/revert.c
  (): Include textbase.h.
  (revert_wc_data): Setaside the text-base contents for the work item.

* subversion/libsvn_wc/update_editor.c
  (update_keywords_after_switch_cb): Setaside the text-base contents for
   the work item.
  (svn_wc_add_repos_file4): Rework to always construct the file_install
   work item with an explicitly provided source file.

* subversion/libsvn_wc/wc_db_update_move.c
  (tc_editor_add_file,
   tc_editor_alter_file,
   tc_editor_update_add_merge_files,
   update_locally_added_node): Setaside the text-base contents for all
   file_install work items.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/pristines-on-demand@1892648 13f79535-47bb-0310-9956-ffa450edef68