On the 'multi-wc-format' branch: include a place-holder format bump.
This patch merely removes the compile-time conditional around the format
bump.  This resolves a compilation error and also makes it easier to merge
with a branch that includes a format bump (such as the 'pristines-on-demand'
branch).

The format bump now included is a no-op. The format number increases from
31 to 32 but nothing about the format is changed, and the client can work
with a WC in either of those formats. This format bump is intended as a
place-holder for some real format change that will be made afterwards.

* subversion/libsvn_wc/wc.h
  (SVN_WC__VERSION): Remove '#ifdef SVN_TEST_MULTI_WC_FORMAT'; bump to 32
    unconditionally.

* subversion/libsvn_wc/upgrade.c
  (bump_to_32,
   svn_wc__version_string_from_format,
   svn_wc__update_schema): Remove '#ifdef SVN_TEST_MULTI_WC_FORMAT'.

* subversion/libsvn_wc/wc-metadata.sql
  Remove '#ifdef SVN_TEST_MULTI_WC_FORMAT', resolving a compile error.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/multi-wc-format@1897698 13f79535-47bb-0310-9956-ffa450edef68