Factor out restore-missing-file code in the WC update code.
No functional change.

There is one minor change. The order of checking disk-kind first followed by
db-kind, in deciding whether to restore the top level node in
svn_wc_crawl_revisions5(), is now reversed, becoming consistent with the
order it uses in checking every other node in the sub-tree in
report_revisions_and_depths() (which is disk-kind then db-kind). This could
in principle make a difference to timings and failure modes.

* subversion/libsvn_wc/adm_crawler.c
  (svn_wc_restore): Move after 'restore_node' and use 'restore_node' rather
    than duplicating that logic.
  (maybe_restore_node): New, factored out...
  (report_revisions_and_depths,
   svn_wc_crawl_revisions5): ... from these.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1898847 13f79535-47bb-0310-9956-ffa450edef68