On the 'pristines-on-demand-on-mwf' branch: Add a new file with explicit test cases for a pristines-on-demand working copy.
I plan to add an option to run the whole test suite in this mode, but this
change ensures that we have a set of tests for the feature that runs by
default.  Also, we'd have a dedicated location to write tests for certain
cases that are specific to the pristines-on-demand mode.

* subversion/tests/cmdline/pristines_on_demand_tests.py: New file, containing…
  (simple_checkout,
   simple_commit,
   simple_update,
   simple_status,
   simple_diff,
   simple_revert,
   update_modified_file): …these new tests.

* subversion/tests/cmdline/svntest/main.py
  (is_wc_pristines_on_demand_supported): New predicate.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/pristines-on-demand-on-mwf@1905384 13f79535-47bb-0310-9956-ffa450edef68