On the 'pristines-on-demand-on-mwf' branch: Skip basic_tests.py#9 when a working copy is created with --store-pristine=no.
The test manually edits mu's text-base when mu is unmodified.  Unmodified
files don't have their text-bases available with --store-pristine=no, so
I don't think we can adjust this test case to run without making significant
changes.  We could add a separate test that would corrupt the text-base of
the file with local mods and run update, but I'd rather keep the original
test case as-is.

* subversion/tests/cmdline/basic_tests.py
  (basic_update_corruption): Remove Wimp() marker.  Skip if the working
   copy is configured not to store local pristines.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/pristines-on-demand-on-mwf@1905570 13f79535-47bb-0310-9956-ffa450edef68