On the 'pristines-on-demand-on-mwf' branch: In the update_tests.py#57, specify and test the current behavior.
For a working copy that doesn't store local pristine contents, we don't
skip access violation errors when determining if a file was modified.
Because if we did that (and treated such files as modified, as we do for
working copies with pristines), we'd also find ourselves uncontrollably
fetching and removing pristines based on transient errors — which probably
is an undesired property.

* subversion/tests/cmdline/update_tests.py
  (skip_access_denied): Remove Wimp() marker.  Expect an error if
   the working copy doesn't store local pristine copies.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/pristines-on-demand-on-mwf@1905601 13f79535-47bb-0310-9956-ffa450edef68