On the 'pristines-on-demand-on-mwf' branch: Skip the revert_reexpand_keyword() test for working copies that don't store local pristines.
The test expects two different things to happen during revert, depending
on whether an `svn cleanup` or any other command that repairs the timestamps
was called in between:

- In the first part, we expect a revert to change file contents and
  re-expand its keywords, because a revert happens right after editing
  the file.

- In the second part, we expect a revert to skip the file with exactly
  the same contents, because there's an in-between operation that has
  recorded the new unmodified timestamp in the db.

This is a problem, because we expect two different things to happen for
two identical on-disk file states.  The only difference is whether any
of the commands that perform the internal timestamp bookkeeping was
called in between.  For example, calling `svn cleanup` after editing the
file makes the first and the second parts of the test behave identically.

This problem currently prevents us from testing the `--store-pristine=no`
case, because when we walk the text-bases we use that opportunity to repair
timestamps, but that also makes the first and the second reverts in this
test behave identically.  So let's postpone testing this case until we
resolve the described problem with opposite expectations.

* subversion/tests/cmdline/revert_tests.py
  (revert_reexpand_keyword): Remove XFail() marker.  Skip for working copies
   that don't store local pristines.  Adjust comments.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/pristines-on-demand-on-mwf@1905756 13f79535-47bb-0310-9956-ffa450edef68