Fix cmdline parsing bug: Add check of the changeno_end variable for zero.
The parser of the --change argument already checks the revision
number for a zero and raises an error, because there are no changes.
However, if a range is given to this argument it would not check
its second part, and the command aborts.

Command to reproduce:
$ svn diff https://svn.apache.org/repos/asf -c 1-0

Adding the check of the changeno_end variable for zero will fix
the problem.

* subversion\svn\svn.c
  (sub_main): Add check of the changeno_end variable for zero.

* subversion\tests\cmdline\diff_tests.py
  (diff_invalid_change_arg): Add test case for a diff of change,
   done in revision range '1-0' and expect an error from it.

Found by: jun66j5

Patch by: Timofey Zhakov (tima {at} chemodax _dot_ net)


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1917944 13f79535-47bb-0310-9956-ffa450edef68