Use svn_opt_parse_change_to_range function to parse the --change argument in the svnbench program.
In r1918076, the svn_opt_parse_change_to_range function was
factored-out from the svn program into the libsvn_subr library.
Since the same functionality is needed for parsing the same argument
in the svnbench program, we can use this function there as well.

Additionally, this refactoring would fix some parsing bugs related to this
argument, which were resolved in the svn program [1], but still exist in the
svnbench program, because this code was duplicated.

* subversion/svnbench/svnbench.c
  (sub_main): Replace the code that parses the --change argument with
   svn_opt_parse_change_to_range function.

[1] The fixes in the --change argument parser were committed in
    r1917944 and r1917864.

Patch by: Timofey Zhakov (tima {at} chemodax _dot_ net)


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1918181 13f79535-47bb-0310-9956-ffa450edef68