On the 'cmake' branch: Link against FS and RA modules.
This is done by first adding the modules to SVN_FS_MODULES or SVN_RA_MODULES
CMake variables. After, when declaring dependencies of a target, `fs-lib`
or `ra-libs` dependencies from build.conf will be replaced to
${SVN_FS_MODULES} or ${SVN_RA_MODULES}. After that, they would be expanded
by CMake.

This makes it possible to disable a specific RA or FS modules, because it
just would not be added to a group, so a library would not link with it.
This feature is not currently implemented.

The `gen_cmake.py` file declares the `group` field in the cmake_target class
and then passes it to the CMakeLists.txt.ezt template file, which generates
a `list(APPEND)` [1] command for the SVN_**_MODULES variables.

It is not required to declare these variable before, because CMake guaranties
that `If no variable named <list> exists [...] the elements are appended to
that empty list` [1].

* build/generator/gen_cmake.py
  (cmake_target): Add a `group` field to the class.

  (Generator.write): Initialize the `group` variable to pass it to the
   cmake_target class.

  (Generator.write, dependencies loop): If `ra-libs` or `fs-libs` are met,
   add ${SVN_**_MODULES} to the libs list.

* build/generator/templates/CMakeLists.txt.ezt
  (): If `group` exists, generate a `list(APPEND)` command, to add target
   name to the specific by that field group.

References:
[1] https://cmake.org/cmake/help/latest/command/list.html#append

git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918736 13f79535-47bb-0310-9956-ffa450edef68