On the 'cmake' branch: Allow specific FS and RA modules to be disabled.
Now, it is possible to specify any of the following options when configuring
CMake. By default, they are set to ON, which means that they should be built,
but if they are set to OFF, this target would not be included to build and
not linked by other libraries.

Here is the list with all new options:

- SVN_BUILD_RA_LOCAL
- SVN_BUILD_RA_SERF
- SVN_BUILD_RA_SVN
- SVN_BUILD_FS_FS
- SVN_BUILD_FS_X

* build/cmake/options.cmake
  (): Add options for building specific FS and RA modules and setup defines
   based on them.

* build/generator/gen_cmake.py
  (get_module_name): New function, for getting the name of the module.
  (Generator.write): Setup enable_condition for FS and RA modules.

* subversion/svn_private_config.hc
  (): Do not setup SVN_LIBSVN_**_LINKS_**_** defines, because they are
   now set by using add_compile_definitions commands, directly from CMake.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/cmake@1918750 13f79535-47bb-0310-9956-ffa450edef68